requirements.txt
Flask==2.0.1
Flask-WTF==1.0.0
Werkzeug==2.0.1
python-dotenv==0.19.0
google-api-python-client==2.36.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
PyPDF2==2.0.0
openpyxl==3.0.9
python-docx==0.8.11
jira==3.1.1
python-dotenv==0.19.0
